added three new tabs on deparment new employee

+++ server/app.py
@@ -48,7 +48,7 @@ class EmployeeResource(Resource):
         return make_response(jsonify(new_employee.to_dict()), 201)
 
     def patch(self, employee_id):
-        employee = Employee.query.filter(Employee.id == employee_id).first()
+        employee = Employee.query.filter(Employee.employee_id == employee_id).first()
         data = request.get_json()
         employee.first_name = data.get("first_name", employee.first_name)
         employee.last_name = data.get("last_name", employee.last_name)
@@ -60,11 +60,15 @@ class EmployeeResource(Resource):
         return make_response(jsonify(employee.to_dict()), 200)
 
     def delete(self, employee_id):
-        employee = Employee.query.filter(Employee.id == employee_id).first()
+        employee = Employee.query.filter(Employee.employee_id == employee_id).first()
+
+        if not employee:
+            return make_response(jsonify({"message": "Employee not found"}), 404)
+
         db.session.delete(employee)
         db.session.commit()
         return make_response(
-            jsonify({"message": f"Employee {employee.id} deleted"}), 204
+            jsonify({"message": f"Employee {employee.employee_id} deleted"}), 200
         )
